refactor(parts): generate json from yaml data
