+++ packages/parts/data/sprite_defs.yaml
