--- packages/parts/src/__mocks__/sprite_defs.yaml
@@ -1,12 +1,0 @@
-foo:
-  id: foo
-  originY: 120
-bar:
-  id: bar
-  offsetX: 45
-baz:
-  id: baz
-  offsetX: -315
-  originY: -160
-qux:
-  id: qux
