+++ packages/parts/src/data/README.md
@@ -1,0 +1,1 @@
+Files in this folder are generated from the `data` folder at the root of this workspace. Do not edit manually!
