+++ packages/parts/src/data/sprite_defs.json
