+++ packages/parts/src/info.test.ts
@@ -1,4 +1,4 @@
-import SPRITE_DEFS from './sprite_defs.json' with { type: 'json' }
+import SPRITE_DEFS from './data/sprite_defs.json' with { type: 'json' }
 import {
   getSpriteDef,
   getAllSegmentInfo,
@@ -6,10 +6,6 @@ import {
   getSegmentVariantInfo,
 } from './info.js'
 
-// vi.mock(
-//   './sprite_defs.yaml',
-//   async () => await import('./__mocks__/sprite_defs.yaml')
-// )
 vi.mock(
   './sprite_defs.json',
   async () => await import('./__mocks__/sprite_defs.json')
